Refactored user controllers.
